Move solve from player to puzzle
